pic_length should return int type

+++ extlib/benz/include/picrin/pair.h
@@ -66,7 +66,7 @@ pic_value pic_make_list(pic_state *, size_t, pic_value);
 #define pic_push(pic, item, place) (place = pic_cons(pic, item, place))
 #define pic_pop(pic, place) (place = pic_cdr(pic, place))
 
-size_t pic_length(pic_state *, pic_value);
+int pic_length(pic_state *, pic_value);
 pic_value pic_reverse(pic_state *, pic_value);
 pic_value pic_append(pic_state *, pic_value, pic_value);
